Refactor feature flag types and imports for improved organization and clarity
